core/requirements.txt
# AI婚恋匹配系统依赖库
# Python >= 3.8

# 核心依赖
requests>=2.28.0          # HTTP请求库（调用通义千问API）
python-dotenv>=0.19.0     # 环境变量配置管理

# 可选依赖（用于增强功能）
# 如需使用文本处理NLP功能，取消注释以下行：
# jieba>=0.42.1           # 中文分词
# paddlenlp>=2.6.0        # PaddleNLP文本处理

# 开发依赖（可选）
# pytest>=7.0             # 单元测试框架
# black>=22.0             # 代码格式化
# flake8>=4.0             # 代码检查